+++ LibraryManagement/src/main/java/app/repository/ReservationRepository.java
@@ -9,8 +9,8 @@ import org.springframework.stereotype.Repository;
 import app.entity.Reservation;
 
 @Repository
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
+public interface ReservationRepository extends JpaRepository<Reservation, Integer> {
 
     @EntityGraph(attributePaths = {"book"})
-    List<Reservation> findByMember_MemberId(Long memberId);
+    List<Reservation> findByMember_MemberId(Integer memberId);
 }
